+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 #### 0.0.x Releases
+- `0.0.23` Releases - [0.0.23](#0023)
 - `0.0.22` Releases - [0.0.22](#0022)
 - `0.0.21` Releases - [0.0.21](#0021)
 - `0.0.20` Releases - [0.0.20](#0020)
@@ -19,6 +20,17 @@ All notable changes to this project will be documented in this file.
 - `0.0.3` Releases - [0.0.3](#003)
 - `0.0.2` Releases - [0.0.2](#002)
 - `0.0.1` Releases - [0.0.1](#001)
+
+## [0.0.23]
+Released on 2020-04-15
+
+#### Added
+- `SwiftPackage` support
+- increase target plataform to 13.0
+
+#### Updated
+- `UIViewController` remove deprecated code for keyWindow.
+
 
 ## [0.0.22]
 Released on 2019-08-07